requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pandas
python-dotenv
